Moving the onboarding fragments in the onboarding package
